client-manager/requirements.txt
fastapi
uvicorn
pymongo
SQLAlchemy
pymysql
python-dotenv
grpcio
grpcio-tools
httpx
pydantic